Fix some case mismatch issue(the naming of source file with CLass name)

+++ app/src/com/outsource/ecg/ui/BaseChartView.java
@@ -508,8 +508,8 @@ public class BaseChartView extends View
      *
      * @return The scale factor.
      */
-    public double getScaleX() {
-        return this.scaleX;
+    public float getScaleX() {
+        return (float) this.scaleX;
     }
 
     /**
@@ -518,8 +518,8 @@ public class BaseChartView extends View
      *
      * @return The scale factor.
      */
-    public double getScaleY() {
-        return this.scaleY;
+    public float getScaleY() {
+        return (float) this.scaleY;
     }
 
     /**
